move navbar and form to app
