+++ src/pages/experience.tsx
@@ -1,0 +1,13 @@
+import { Navbar } from "@/components";
+
+const Experience = () => {
+  return (
+    <div>
+      <p className="flex grow flex-col justify-start text-3xl text-white">
+        Currently WIP
+      </p>
+    </div>
+  );
+};
+
+export default Experience;
