+++ src/pages/projects.tsx
@@ -1,0 +1,8 @@
+const Projects = () => {
+    return <div>
+            <p className="text-white text-3xl">Currently WIP</p>
+    </div>  
+    
+}
+
+export default Projects;
